Rewrite all docs for new examples and better way to handle it
