+++ docs/App.js
@@ -1,19 +1,13 @@
-import React, { Component } from 'react'
+import React, { Component, Fragment } from 'react'
 
 import Lowlight from 'react-lowlight'
 import js from 'highlight.js/lib/languages/javascript'
 import 'highlight.js/styles/monokai-sublime.css'
 
+import pages from './pages/*.js'
+
 // Then register them with lowlight
 Lowlight.registerLanguage('js', js)
-
-import pages from './pages/*.jsx';
-
-const navButtonsContent = `
-<a class="github-button" href="https://github.com/miduga/react-slidy" aria-label="React Slidy Repo">Repo</a>
-<a class="github-button" href="https://github.com/miduga/react-slidy" data-icon="octicon-star" data-count-href="/miduga/react-slidy/stargazers" data-show-count="true" data-count-aria-label="# stargazers on GitHub" aria-label="Star miduga/react-slidy on GitHub">Star</a>
-<a class="github-button" href="https://github.com/miduga/react-slidy/issues" data-icon="octicon-issue-opened" data-show-count="true" data-count-aria-label="# issues on GitHub" aria-label="Issue miduga/react-slidy on GitHub">Issue</a>
-`
 
 export default class App extends Component {
   constructor (...args) {
@@ -26,12 +20,6 @@ export default class App extends Component {
     if (hash) {
       this.setState({ activeSection: hash.replace('#', '') })
     }
-
-    const script = document.createElement('script')
-    script.src = 'https://buttons.github.io/buttons.js'
-    script.async = true
-
-    document.body.appendChild(script)
   }
 
   _changeActivatedSection (demo) {
@@ -44,13 +32,8 @@ export default class App extends Component {
     const {activeSection} = this.state
 
     return (
-      <div>
-        <nav className='pt2'>
-          <h1 className='h1 mb0 mt0'>React Slidy</h1>
-          <div
-            className='nav-buttons mt1 mb2'
-            dangerouslySetInnerHTML={{ __html: navButtonsContent }}
-          />
+      <Fragment>
+        <nav>
           {Object.keys(pages).map(demo => (
             <a
               href={`#${demo}`}
@@ -71,7 +54,7 @@ export default class App extends Component {
           ))}
         </section>
 
-      </div>
+      </Fragment>
     )
   }
 }
