+++ docs/index.html
@@ -3,6 +3,16 @@
 <meta name="viewport" content="width=device-width,user-scalable=no">
 
 <body>
+  <header class='text-center'>
+    <h1>React Slidy 🍃</h1>
+    <h2>a simple and minimal slider component for React focused on performance</h2>
+
+    <div class='github'>
+      <a class="github-button" href="https://github.com/miduga/react-slidy" data-icon="octicon-star" data-size="large" data-show-count="true" aria-label="Star miduga/react-slidy on GitHub">Star project</a>
+      <a class="github-button" href="https://github.com/miduga/react-slidy/issues" data-icon="octicon-issue-opened" data-size="large" data-show-count="true" aria-label="Issue miduga/react-slidy on GitHub">Issue</a>
+    </div>
+    <p>How to use: Just wrap the elements that you want to use on your slider. React Slidy will do the magic for you.</p>
+  </header>
   <div id='root'></div>
 </body>
 
@@ -11,3 +21,6 @@
 
 <link rel="stylesheet" href="../src/index.scss">
 <script src='../src/index.js'></script>
+
+<!-- Place this tag in your head or just before your close body tag. -->
+<script async defer src="https://buttons.github.io/buttons.js"></script>
