+++ docs/pages/with-dynamic-content.js
@@ -39,7 +39,6 @@ class DynamicReactSlidyContent extends Component {
         </button>
         <ReactSlidy
           dynamicContent
-          infinite={false}
           doAfterDestroy={() => {
             console.log('slider destroyed!')
           }}
