+++ docs/pages/with-events.js
