+++ docs/pages/with-initial-index.js
