+++ docs/pages/with-other-components.js
